requirements.txt
asn1crypto==0.24.0
ConfigArgParse==1.2.3
cryptography==2.1.4
enum34==1.1.6
faiss-cpu
idna==2.6
ipaddress==1.0.17
kafka-python==2.0.2
keyring==10.6.0
keyrings.alt==3.0
Logbook==1.5.3
numpy==1.16.6
pandas==0.24.2
psycopg2==2.8.6
pycrypto==2.6.1
python-dateutil==2.8.1
pytz==2020.4
pyxdg==0.25
SecretStorage==2.3.1
six==1.11.0
SQLAlchemy==1.3.20

